test(disaggregation): consolidate PD tests into 8 test_pd_* files
